make sure everything happens in the right order when reconnecting.
previously a login and/or subscription could happen before the connect method call
was sent, sometimes resulting in errors, because the 'reconnected' event was fired
as soon as the socket was connected rather than when a reply was received from the
'connect' ddp call.

also the 'connected' event was fired for reconnects as well as the 'reconnected'
event. this patch makes sure only one or the the other event is fired so you can
really tell whether you just connected or reconnected.