add support for DDP version 1.
need to implement fallback to older versions (PRE1, PRE2) if
connection to old meteor servers is needed.

randomSeed is not implemented as it is not really documented and
seems to be completely optional.

basically all this does is change the version number and add
support for responding to ping messages. everything should be
backwards compatible.